Create voting
Tech Stacks- Solidity, Remix IDE
Project Description: First we choose an address to be our commissioner, then we copy other addresses and use it to register the candidates. A candidate can only be registered by the comissioner.
Now, to vote someone we need to input their address but before that we need to check if it is registered or not. The result setter will calculate the votes and winner shows the winning candidate. The countVote getter can show the number of votes for the corresponding address inputted. Similar is the function of voterecieved. The candidate getter shows the corresponding candidate for the inputed index. The totalvote getter shows the votes receieved by the winner.